more changes acc to review comments bumped up versions
