+++ ai-services/assets/applications/rag-cpu/podman/values.yaml
@@ -85,7 +85,7 @@ summarize:
   # @description Host port for the Summarize API. If unspecified, a random available port is assigned. Specify a port number to use a custom value.
   port: ""
   # @hidden
-  image: icr.io/ai-services-cicd/summarize-service:v0.0.4
+  image: icr.io/ai-services-cicd/summarize-service:v0.0.5
   # @hidden
   log_level: "INFO"
   database: "summarize_metadata"
