+++ ai-services/assets/applications/rag-dev/openshift/values.yaml
@@ -12,7 +12,7 @@ backend:
 
 summarize:
   # @hidden
-  image: icr.io/ai-services-cicd/summarize-service:v0.0.4
+  image: icr.io/ai-services-cicd/summarize-service:v0.0.5
   # @hidden
   log_level: "INFO"
