+++ services/summarize/Makefile
@@ -1,5 +1,5 @@
 IMAGE=summarize-service
-TAG?=v0.0.4
+TAG?=v0.0.5
 
 run:
 	PORT=$${PORT:-6000} /var/venv/bin/python -m uvicorn app:app --host 0.0.0.0 --port $$PORT
